+++ src/PS.FarNet.Redis/GetSetCommand.cs
@@ -1,4 +1,5 @@
 ﻿using StackExchange.Redis;
+using System.Collections.Generic;
 using System.Management.Automation;
 
 namespace PS.FarNet.Redis;
@@ -7,6 +8,13 @@ namespace PS.FarNet.Redis;
 [OutputType(typeof(long))]
 public sealed class GetSetCommand : BaseGetCountCmdlet
 {
+    const string NPattern = "Pattern";
+
+    [Parameter(ParameterSetName = NPattern, Mandatory = true)]
+    [AllowEmptyString]
+    [AllowNull]
+    public string Pattern { get; set; }
+
     protected override void BeginProcessing()
     {
         base.BeginProcessing();
@@ -17,6 +25,16 @@ public sealed class GetSetCommand : BaseGetCountCmdlet
                 {
                     long res = Database.SetLength(RKey);
                     WriteObject(res);
+                }
+                break;
+            case NPattern:
+                {
+                    if (Pattern is { })
+                        Pattern = ConvertPatternToRedis(Pattern);
+
+                    IEnumerable<RedisValue> res = Database.SetScan(RKey, RedisValue.Unbox(Pattern), 250, CommandFlags.None);
+                    foreach (var value in res)
+                        WriteObject((string)value);
                 }
                 break;
             default: